pyproject.toml
[tool.poetry]
name = "coffee-scale-dashboard"
version = "0.1.0"
description = ""
authors = ["shin5819 <shin5819@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dash = "^2.17.1"
dash-bootstrap-components = "^1.6.0"
paho-mqtt = "^2.1.0"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
